Fibonacci heap class

Implementation of a Fibonacci heap data structure, i.e. a priority datastructure with push in amortized O(1) and pop in O(log n). fibonacci_heap wraps a boost::fibonacci_heap using Rcpp modules. The heap consists of nodes with keys and values where the key determines the priority in the heap. Also see the binomial_heap class.

Slots

.heap

C++ object representing a heap

.key.class

the class of the keys