Add elements to an object

Adds keys or <key, value> pairs to an object and returns the object. Depending on the datastructure used, either only keys are required or pairs of <keys, values>. Insertion of elements with vectors, i.e. giving multiple arguments at the same time is faster than inserting elements iteratively.

insert(obj, x, y)

# S4 method for deque,ANY,missing
insert(obj, x)

# S4 method for deque,list,missing
insert(obj, x)

# S4 method for heap,vector,vector
insert(obj, x, y)

# S4 method for heap,vector,matrix
insert(obj, x, y)

# S4 method for heap,vector,list
insert(obj, x, y)

# S4 method for heap,vector,ANY
insert(obj, x, y)

# S4 method for bimap,vector,vector
insert(obj, x, y)

# S4 method for unordered_map,vector,vector
insert(obj, x, y)

# S4 method for unordered_map,vector,list
insert(obj, x, y)

# S4 method for unordered_map,vector,ANY
insert(obj, x, y)

Arguments

obj
object to insert into

x
the values/keys to insert into

y
values to be inserted which are required for some datastructures

Value

returns obj with inserted values

Examples

 # inserts values into a multimap
 m_map <- multimap()
 m_map <- insert(m_map, c("a", "b"), 1:2)
 m_map <- insert(m_map, c("a", "b"), list(1, list(a=1)))
 m_map["a"] <- rnorm(length(letters))
 m_map[c("a", "b", "c")] <- list(1, data.frame(a=2), environment())

 # inserts values into a fibonacci_heap
 f_heap <- fibonacci_heap("integer")
 f_heap <- insert(f_heap, 1:2, 1:2)
 f_heap[3:4] <- list(1, list(a=1))
 f_heap <- insert(f_heap, 5:6, list(data.frame(a=rnorm(3)), diag(2)))

 # inserts elements into a queue or stack
 s <- stack()
 s <- insert(s, list(1, vector(), list(3), data.frame(rnorm(3))))