Pop a single element from an object

Remove and return the first element from a data structure that has a priority, such as a heap or deque.

pop(obj)

# S4 method for deque
pop(obj)

# S4 method for heap
pop(obj)

Arguments

obj
the object to pop an element from

Value

returns the first element from obj as list

Examples

 # pops from a queue
 q <- queue()
 q <- insert(q, list(environment(), data.frame(a=1)))
 pop(q)

#> <environment: 0x56135b82aa10>

 # pops from a stack
 s <- stack()
 s <- insert(s, list(environment(), data.frame(a=1)))
 pop(s)

#>   a
#> 1 1

 # pops from a fibonacci heap
 b_heap <- binomial_heap()
 b_heap <- insert(b_heap, letters[seq(3)], list(1, diag(3), rnorm(2)))
 pop(b_heap)

#> $a
#> [1] 1
#>